+++ server/src/main/java/org/apache/druid/server/http/SegmentLoadingMode.java
@@ -19,6 +19,8 @@
 
 package org.apache.druid.server.http;
 
-/**
- *
- */
+public enum SegmentLoadingMode
+{
+  NORMAL,
+  TURBO;
+}
